Linux build fixes (#70)
* Fix warning and default mono space font when build is not windows or mac
* Update workflow to build for linux

+++ ORchestra/src/ORchestraEngine/MidiScheduler.cpp
@@ -120,7 +120,7 @@ namespace ORchestra {
             }
         }
 
-        mScheduledMidiMessages.erase(mScheduledMidiMessages.begin() + keepCount, mScheduledMidiMessages.end());
+        mScheduledMidiMessages.erase(mScheduledMidiMessages.begin() + static_cast<long>(keepCount), mScheduledMidiMessages.end());
     }
 
     void MidiScheduler::ClearAllData(juce::MidiBuffer& midiMessages)
